requirements.txt
clickhouse-driver==0.2.9
numpy==2.3.1
pandas==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
tzlocal==5.3.1
